Create the pyproject.toml file
This update includes the package install requirements. This update does not include the castor specific packages. The poetry lock file has not been created yet. There was issues with the package "arviz" where it didn't agree with the python version implemented. Other packages after "arviz" may have issues as it didn't load after that package in the list of packages.