Read the Docs: First iteration (#41)
* added docs to gitignore file for now

* started initial draft of ETC auto-generated docs

* added new pages ; fixed some dependency issues to enable markdown parsing

* Added additional extensions to assist with parsing and editing

* added a pre-build script to integrate notebooks into examples

* added some introduction files

* added intro files actually

* Removed _static HTML file for now

* Create the pyproject.toml file

This update includes the package install requirements. This update does not include the castor specific packages. The poetry lock file has not been created yet. There was issues with the package "arviz" where it didn't agree with the python version implemented. Other packages after "arviz" may have issues as it didn't load after that package in the list of packages.

* refactored based on Hatch ; added GH actions

* Added to the introduction

Added text to the introduction for FORECASTOR, the different objects with descriptions and made a simple diagram to show how the objects connect to the Photometry Object.

* refactored documentation folder structures

* changed a bunch of folder structure for documentation; missing content inside for now

* refactored documentation paths

---------

Co-authored-by: Jocelyn <jgroney1@gmail.com>